HUM-87 - Available jobs in worker module refactor
