+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs/available-jobs/components/available-jobs-reward-amount-sort.tsx
