+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs/available-jobs/components/mobile/available-jobs-assign-job-button-mobile.tsx
@@ -6,7 +6,7 @@ import {
 import { TableButton } from '@/shared/components/ui/table-button';
 import { useJobsNotifications } from '@/modules/worker/hooks/use-jobs-notifications';
 
-export function AvailableJobsAssignJobButton({
+export function AvailableJobsAssignJobButtonMobile({
   assignJobPayload,
 }: {
   assignJobPayload: AssignJobBody;
